stt_dev/requirements.txt
faster-whisper
python-dotenv
openai-whisper
python-telegram-bot